+++ sdk/src/androidTest/java/ly/count/android/sdk/ModuleViewsTests.java
@@ -1855,6 +1855,113 @@ public class ModuleViewsTests {
         validateView("test2", 0.0, 1, 2, false, true, TestUtils.map(), "_CLY_", "_CLY_", "test");
     }
 
+    /**
+     * "startView" with consent removal
+     * Validate that all running views are stopped when the view consent is removed
+     *
+     * @throws JSONException if the JSON is not valid
+     */
+    @Test
+    public void startView_consentRemoval() throws JSONException {
+        CountlyConfig countlyConfig = TestUtils.createBaseConfig();
+        countlyConfig.setRequiresConsent(true);
+        countlyConfig.setLoggingEnabled(true);
+        countlyConfig.giveAllConsents();
+        countlyConfig.setEventQueueSizeToSend(1);
+
+        Countly countly = new Countly().init(countlyConfig);
+
+        countly.views().startView("test");
+        ModuleConsentTests.validateAllConsentRequest(TestUtils.commonDeviceId, 0);
+        validateView("test", 0.0, 1, 2, true, true, TestUtils.map(), "_CLY_", "_CLY_", null);
+
+        countly.views().startView("test2");
+        validateView("test2", 0.0, 2, 3, false, true, TestUtils.map(), "_CLY_", "_CLY_", null);
+
+        countly.consent().removeConsent(new String[] { Countly.CountlyFeatureNames.views });
+        validateView("test", 0.0, 3, 6, false, false, TestUtils.map(), "_CLY_", "_CLY_", null);
+        validateView("test2", 0.0, 4, 6, false, false, TestUtils.map(), "_CLY_", "_CLY_", null);
+        ModuleConsentTests.validateConsentRequest(TestUtils.commonDeviceId, 5, new boolean[] { true, true, true, true, true, true, true, true, true, true, true, true, false, true, true });
+
+        countly.consent().giveConsent(new String[] { Countly.CountlyFeatureNames.views });
+        ModuleConsentTests.validateAllConsentRequest(TestUtils.commonDeviceId, 6);
+        Assert.assertEquals(7, TestUtils.getCurrentRQ().length);
+    }
+
+    /**
+     * "startAutoStoppedView" with consent removal
+     * Validate that running view is stopped when the view consent is removed
+     *
+     * @throws JSONException if the JSON is not valid
+     */
+    @Test
+    public void startAutoStoppedView_consentRemoval() throws JSONException {
+        CountlyConfig countlyConfig = TestUtils.createBaseConfig();
+        countlyConfig.setRequiresConsent(true);
+        countlyConfig.setLoggingEnabled(true);
+        countlyConfig.giveAllConsents();
+        countlyConfig.setEventQueueSizeToSend(1);
+
+        Countly countly = new Countly().init(countlyConfig);
+
+        countly.views().startAutoStoppedView("test");
+        ModuleConsentTests.validateAllConsentRequest(TestUtils.commonDeviceId, 0);
+        validateView("test", 0.0, 1, 2, true, true, TestUtils.map(), "_CLY_", "_CLY_", null);
+
+        countly.views().startAutoStoppedView("test2");
+        validateView("test", 0.0, 2, 4, false, false, TestUtils.map(), "_CLY_", "_CLY_", null);
+        validateView("test2", 0.0, 3, 4, false, true, TestUtils.map(), "_CLY_", "_CLY_", null);
+
+        countly.consent().removeConsent(new String[] { Countly.CountlyFeatureNames.views });
+        validateView("test2", 0.0, 4, 6, false, false, TestUtils.map(), "_CLY_", "_CLY_", null);
+        ModuleConsentTests.validateConsentRequest(TestUtils.commonDeviceId, 5, new boolean[] { true, true, true, true, true, true, true, true, true, true, true, true, false, true, true });
+
+        countly.consent().giveConsent(new String[] { Countly.CountlyFeatureNames.views });
+        ModuleConsentTests.validateAllConsentRequest(TestUtils.commonDeviceId, 6);
+        Assert.assertEquals(7, TestUtils.getCurrentRQ().length);
+    }
+
+    /**
+     * Auto view tracking with consent removal
+     * Validate that running view is stopped when the view consent is removed
+     *
+     * @throws JSONException if the JSON is not valid
+     */
+    @Test
+    public void autoViewTracking_consentRemoval() throws JSONException {
+        CountlyConfig countlyConfig = TestUtils.createBaseConfig(TestUtils.getContext());
+        countlyConfig.setRequiresConsent(true);
+        countlyConfig.setLoggingEnabled(true);
+        countlyConfig.enableAutomaticViewTracking();
+        countlyConfig.giveAllConsents();
+        countlyConfig.setEventQueueSizeToSend(1);
+
+        Countly countly = new Countly().init(countlyConfig);
+
+        Activity activity = Mockito.mock(Activity.class);
+        countly.onStart(activity);
+
+        ModuleConsentTests.validateAllConsentRequest(TestUtils.commonDeviceId, 0);
+        ModuleSessionsTests.validateSessionBeginRequest(1, TestUtils.commonDeviceId);
+        ModuleEventsTests.validateEventInRQ("[CLY]_orientation", TestUtils.map("mode", "portrait"), 1, 0, 0, 2, 4);
+
+        validateView(activity.getClass().getName(), 0.0, 3, 4, true, true, TestUtils.map(), "_CLY_", "_CLY_", null);
+
+        Activity activity2 = Mockito.mock(Activity.class);
+        countly.onStart(activity2);
+        validateView(activity.getClass().getName(), 0.0, 4, 6, false, false, TestUtils.map(), "_CLY_", "_CLY_", null);
+        validateView(activity2.getClass().getName(), 0.0, 5, 6, false, true, TestUtils.map(), "_CLY_", "_CLY_", null);
+
+        countly.consent().removeConsent(new String[] { Countly.CountlyFeatureNames.views });
+        validateView(activity2.getClass().getName(), 0.0, 6, 8, false, false, TestUtils.map(), "_CLY_", "_CLY_", null);
+        ModuleConsentTests.validateConsentRequest(TestUtils.commonDeviceId, 7, new boolean[] { true, true, true, true, true, true, true, true, true, true, true, true, false, true, true });
+
+        countly.consent().giveConsent(new String[] { Countly.CountlyFeatureNames.views });
+
+        ModuleConsentTests.validateAllConsentRequest(TestUtils.commonDeviceId, 8);
+        Assert.assertEquals(9, TestUtils.getCurrentRQ().length);
+    }
+
     static void validateView(String viewName, Double viewDuration, int idx, int size, boolean start, boolean visit, Map<String, Object> customSegmentation, String id, String pvid) throws JSONException {
         validateView(viewName, viewDuration, idx, size, start, visit, customSegmentation, id, pvid, null);
     }
@@ -1877,6 +1984,5 @@ public class ModuleViewsTests {
 
         ModuleEventsTests.validateEventInRQ(TestUtils.commonDeviceId, ModuleViews.VIEW_EVENT_KEY, viewSegmentation, 1, 0.0, viewDuration, id, pvid, "_CLY_", "_CLY_", idx, size, 0, 1);
     }
-
     //todo extract orientation tests
 }